requirements.txt
Django==2.1
djangorestframework==3.8.2
psycopg2==2.7.2
django-cors-headers==2.4.0
dj-database-url==0.4.2
gunicorn==19.7.1
virtualenv==15.1.0
requests==2.19.1
